services/tts/requirements.txt
chatterbox-tts
numpy==1.26.0
Flask
gunicorn
